// blitzkarte/models/enumeration/order-enums.ts
export enum OrderDisplay {
  AIRLIFT = 'Airlift',
  CONVOY = 'Convoy',
  DISBAND = 'Disband',
  HOLD = 'Hold',
  MOVE = 'Move',
  MOVE_CONVOYED = 'Move Convoyed',
  NUKE = 'Nuke',
  OFFER_TECH = 'Offer Nuke Tech',
  RECEIVE_TECH = 'Receive Nuke Tech From',
  RUSH_BUILD = 'Rush Build',
  SUPPORT = 'Support',
  SUPPORT_CONVOYED = 'Support Convoyed',
  TRANSFER = 'Transfer To'
}